chore(wsgi): remove asm context code from wsgi contrib (#6331)
This change removes the use of `asm_request_context_manager` from the
wsgi contrib, making separation of concerns clearer.

In the case of request blocking, the separation of concerns ideally
breaks down as follows. The AppSec Product code in the `ddtrace/appsec`
directory knows how to make a block/don't block decision based on
communication with libddwaf. The Wsgi code in `ddtrace/contrib` knows
how to take that blocking decision into account when processing
requests.