fix(lib-injection): support non-root users (#6047)
The current mechanism attempts a `pip install` using the current user.
Many users run their containers with a non-root user (`USER 1000` or
similar in their Dockerfiles) for security purposes. The injection
mechanism doesn't work with this as it requires modifying the
site-packages in the application container to install the ddtrace
package which most often requires super user permissions.

A solution is to install the package to the volume mount in which the
wheels are provided in a custom `site-packages` directory. The volume
mount should have read/write permissions for any user. This custom
`site-packages` is then included in the PYTHONPATH to make it visible to
the application.

A reproduction of the issue is provided with
[49203e6](https://github.com/DataDog/dd-trace-py/pull/6047/commits/49203e6a9dbf0d9bd092882bcbdf046820657559).
The simple django test image is modified to reproduce the issue and
assert that the problem is fixed.

- [ ] What happens when you pip install with existing deps
- [x] What happens when ddtrace-run is already used in the app? The
ddtrace that is preinstalled should be used.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [ ] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.

---------

Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>