feat(profiling): provide libdatadog as an export method (#5969)
Builds off of #5967. Operationally, it might also depend on #5968
because bringing libdatadog as a dependency costs about ~3 megabytes of
storage, whereas #5968 reduces overhead enough to account for this (and
more!).

This adds support for using libdatadog as an optional backend for the
profiler. This decreases memory and increases performance, but it isn't
1-1 complete with the Python profiler yet.

To enable libdatadog mode
DD_PROFILING_EXPORT_LIBDD_ENABLE=true

To disable the Python profiler
DD_PROFILING_EXPORT_PY_ENABLE=false

It is possible for both exporters to be active and when that happens
libdd annotates its profile version string accordingly (this is useful
for comparison).

The expectation is that One Day libdd will replace the py implementation
entirely and we can remove defunct code, but before then we have a
significant testing gap to overcome. I've been using "dual mode" to
ensure libdatadog has identical output to py, which is fine as far as
manual workflows go, but since libdatadog doesn't provide the degree of
introspection the current collector does, we'll take some effort to use
a comparable mechanism to reimplement correctness tests.

**note**: this is an experimental feature.  Testing will be brought to parity with the existing collector/exporter over subsequent PRs.

## Checklist

- [X] Change(s) are motivated and described in the PR description.
- [X] Testing strategy is described if automated tests are not included
in the PR.
- [X] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [X] Change is maintainable (easy to change, telemetry, documentation).
- [X] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [X] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.

---------

Co-authored-by: sanchda <sanchda@users.noreply.github.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>