Try to make exits safer
During execution of `atexit()` in a django/gunicorn app, I've seen
_schedler.join() throw an exception within threading.py.  I suspect
this has something to do with a rare combination of factors leading to
a bad order in cleanup operations, but I'm not sure.

Because this is happening in `atexit()`, I presume that the consequences
to the end-user are absolutely minimal, but proposing this change in
order to surface a little bit more information.